Project2/testing/ElectionTypeTest.java
import org.junit.*;
import java.util.ArrayList;
import java.util.Arrays;
import static org.junit.jupiter.api.Assert.*;


public class ElectionTypeTest {

    @Test
    public void testCreateCandidateList() {
        ElectionType electionType = new ElectionType() {
            public void runElection() {
                // Implementation not required for this test
            }
        };
        String candidateEntries = "Rosen D, Kleinberg R, Chou I, Royce L";
        ArrayList<Candidate> candidates = electionType.createCandidateList(candidateEntries);

        Assert.assertEquals(4, candidates.size());
        Assert.assertEquals("Rosen", candidates.get(0).getName());
        Assert.assertEquals("D", candidates.get(0).getCandidateParty().getName());
        Assert.assertEquals("Kleinberg", candidates.get(1).getName());
        Assert.assertEquals("R", candidates.get(1).getCandidateParty().getName());
        Assert.assertEquals("Chou", candidates.get(2).getName());
        Assert.assertEquals("I", candidates.get(2).getCandidateParty().getName());
        Assert.assertEquals("Royce", candidates.get(3).getName());
        Assert.assertEquals("L", candidates.get(3).getCandidateParty().getName());
        // More assertions can be added for other candidates
    }


    @Test
    public void testTieBreakerEqualChance() {
        ElectionType electionType = new ElectionType() {
            public void runElection() {
                // Implementation not required for this test
            }
        };
        Candidate candidate1 = new Candidate("Candidate1");
        Candidate candidate2 = new Candidate("Candidate2");

        int winCount1 = 0, winCount2 = 0;
        for (int i = 0; i < 1000; i++) {
            Candidate winner = (Candidate) electionType.tieBreaker(new ArrayList<>(Arrays.asList(candidate1, candidate2)));
            if (winner.equals(candidate1)) winCount1++;
            if (winner.equals(candidate2)) winCount2++;
        }

        Assert.assertTrue(winCount1 >= 450 && winCount1 <= 550); // Check if winCount1 is within 45% - 55%
        Assert.assertTrue(winCount2 >= 450 && winCount2 <= 550); // Check if winCount2 is within 45% - 55%
    }


    @Test
    public void testTieBreakerUnequalChance() {
        ElectionType electionType = new ElectionType() {
            public void runElection() {
                // Implementation not required for this test
            }
        };
        Candidate candidate1 = new Candidate("Candidate1");
        Candidate candidate2 = new Candidate("Candidate2");
        Candidate candidate3 = new Candidate("Candidate3");

        int winCount1 = 0, winCount2 = 0, winCount3 = 0;
        for (int i = 0; i < 1000; i++) {
            Candidate winner = (Candidate) electionType.tieBreaker(new ArrayList<>(Arrays.asList(candidate1, candidate2, candidate3)));
            if (winner.equals(candidate1)) winCount1++;
            if (winner.equals(candidate2)) winCount2++;
            if (winner.equals(candidate3)) winCount3++;
        }

        Assert.assertTrue(winCount1 >= 280 && winCount1 <= 370); // Check if winCount1 is within 28% - 37%
        Assert.assertTrue(winCount2 >= 280 && winCount2 <= 370); // Check if winCount2 is within 28% - 37%
        Assert.assertTrue(winCount3 >= 280 && winCount3 <= 370); // Check if winCount3 is within 28% - 37%
    }


    @Test
    public void testGettersAndSetters() {
        ElectionType electionType = new ElectionType() {
            public void runElection() {
                // Implementation not required for this test
            }
        };

        electionType.setBallotCount(10);
        electionType.setCandidatesCount(4);
        electionType.setElectionType("IR");

        Assert.assertEquals(10, electionType.getBallotCount());
        Assert.assertEquals(4, electionType.getCandidatesCount());
        Assert.assertEquals("IR", electionType.getElectionType());
    }
}
